Refactoring
Refactored reducers into separate folder
