+++ my-app/src/context/product-listing-context.js
@@ -1,23 +1,7 @@
 import React, { createContext, useContext, useReducer } from 'react'
+import productListReducer from './reducers/productListReducer'
 
 const ProductListingContext = createContext()
-
-const productListReducer = (state,{type,payload}) => {
-    console.log("Type",type,"Payload",payload)
-    switch(type){
-        case "SORTING":
-            return ({...state,sorting : payload})
-        case "CATEGORY":
-            return ({...state , category : {...state['category'] , [payload] : !state.category[payload]} })
-        case "RATING":
-            return ({...state,rating : payload})
-        case "RANGE":
-            return ({...state,priceRange: payload})
-        default:
-            break;
-    }
-}
-
 
 const ProductListingContextProvider = ({children}) => {
     const [state , dispatch] = useReducer(productListReducer,{category : {fiction : true , 'non-fiction' : true , horror: true} , rating : "" , sorting : "" , priceRange : 1000})
